Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,9 +1,9 @@
 #include<stdio.h>
 int main(){
-    float a,b;
+    int a,b;
     char c;
-    scanf("%f",&a);
-    scanf("%f",&b);
+    scanf("%i",&a);
+    scanf("%i",&b);
     scanf(" %c",&c);
     int x = a*b;
     int y = a/b;
